+++ include/Sound.h
@@ -43,7 +43,7 @@ public:
      * \brief Get number of registered sources
      * \returns number of registered sources
      */
-    virtual int registeredSources() = 0;
+    virtual unsigned long registeredSources() = 0;
 
     /*!
      * \brief Remove all loaded sounds
@@ -62,7 +62,7 @@ public:
      * \param source valid source id
      * \param pos new position for source
      */
-    virtual void sourceAt(int source, float pos[3]) = 0;
+    virtual void sourceAt(unsigned long source, float pos[3]) = 0;
 
     /*!
      * \brief Load wav file from disk
@@ -71,7 +71,7 @@ public:
      * \param flags set options. Use SoundFlags enum bitwise OR-ed
      * \returns 0 for no error or < 0 error flag
      */
-    virtual int addFile(const char *filename, int *source, unsigned int flags) = 0;
+    virtual int addFile(const char *filename, unsigned long *source, unsigned int flags) = 0;
 
     /*!
      * \brief Load wav file from buffer
@@ -81,19 +81,19 @@ public:
      * \param flags set options. Use SoundFlags enum bitwise OR-ed
      * \returns 0 for no error or < 0 error flag
      */
-    virtual int addWave(unsigned char *wav, unsigned int length, int *source, unsigned int flags) = 0;
+    virtual int addWave(unsigned char *wav, unsigned int length, unsigned long *source, unsigned int flags) = 0;
 
     /*!
      * \brief Play sound source
      * \param source sound source to play
      */
-    virtual void play(int source) = 0;
+    virtual void play(unsigned long source) = 0;
 
     /*!
      * \brief Stop playing sound source
      * \param source sound source to stop
      */
-    virtual void stop(int source) = 0;
+    virtual void stop(unsigned long source) = 0;
 };
 
 Sound &getSound();
