+++ src/World.cpp
@@ -19,11 +19,11 @@ void World::addRoom(Room &room) {
     mRooms.push_back(&room);
 }
 
-unsigned int World::sizeRoom() {
+unsigned long World::sizeRoom() {
     return mRooms.size();
 }
 
-Room &World::getRoom(unsigned int index) {
+Room &World::getRoom(unsigned long index) {
     assert(index < mRooms.size());
     return *mRooms.at(index);
 }
@@ -32,11 +32,11 @@ void World::addSprite(SpriteSequence &sprite) {
     mSprites.push_back(&sprite);
 }
 
-unsigned int World::sizeSprite() {
+unsigned long World::sizeSprite() {
     return mSprites.size();
 }
 
-SpriteSequence &World::getSprite(unsigned int index) {
+SpriteSequence &World::getSprite(unsigned long index) {
     assert(index < mSprites.size());
     return *mSprites.at(index);
 }
@@ -45,11 +45,11 @@ void World::addEntity(Entity &entity) {
     mEntities.push_back(&entity);
 }
 
-unsigned int World::sizeEntity() {
+unsigned long World::sizeEntity() {
     return mEntities.size();
 }
 
-Entity &World::getEntity(unsigned int index) {
+Entity &World::getEntity(unsigned long index) {
     assert(index < mEntities.size());
     return *mEntities.at(index);
 }
@@ -58,11 +58,11 @@ void World::addSkeletalModel(SkeletalModel &model) {
     mModels.push_back(&model);
 }
 
-unsigned int World::sizeSkeletalModel() {
+unsigned long World::sizeSkeletalModel() {
     return mModels.size();
 }
 
-SkeletalModel &World::getSkeletalModel(unsigned int index) {
+SkeletalModel &World::getSkeletalModel(unsigned long index) {
     assert(index < mModels.size());
     return *mModels.at(index);
 }
@@ -71,20 +71,20 @@ void World::addStaticMesh(StaticMesh &model) {
     mMeshes.push_back(&model);
 }
 
-unsigned int World::sizeStaticMesh() {
+unsigned long World::sizeStaticMesh() {
     return mMeshes.size();
 }
 
-StaticMesh &World::getStaticMesh(unsigned int index) {
+StaticMesh &World::getStaticMesh(unsigned long index) {
     assert(index < mMeshes.size());
     return *mMeshes.at(index);
 }
 
 
-int World::getRoomByLocation(int index, float x, float y, float z)
+long World::getRoomByLocation(long index, float x, float y, float z)
 {
     assert(index >= 0);
-    assert(index < (int)mRooms.size());
+    assert(index < (long)mRooms.size());
     Room &room = *mRooms.at(index);
 
     if (room.getBoundingBox().inBox(x, y, z))
@@ -94,10 +94,10 @@ int World::getRoomByLocation(int index, float x, float y, float z)
 }
 
 
-int World::getRoomByLocation(float x, float y, float z) {
-    int hop = -1;
+long World::getRoomByLocation(float x, float y, float z) {
+    long hop = -1;
 
-    for (unsigned int i = 0; i < mRooms.size(); i++) {
+    for (unsigned long i = 0; i < mRooms.size(); i++) {
         if (mRooms.at(i)->getBoundingBox().inBoxPlane(x, z)) {
             if (mRooms.at(i)->getBoundingBox().inBox(x, y, z))
                 return i;
@@ -111,23 +111,23 @@ int World::getRoomByLocation(float x, float y, float z) {
 
 
 void World::destroy() {
-    for (unsigned int i = 0; i < mRooms.size(); i++)
+    for (unsigned long i = 0; i < mRooms.size(); i++)
         delete mRooms[i];
     mRooms.clear();
 
-    for (unsigned int i = 0; i < mSprites.size(); i++)
+    for (unsigned long i = 0; i < mSprites.size(); i++)
         delete mSprites[i];
     mSprites.clear();
 
-    for (unsigned int i = 0; i < mEntities.size(); i++)
+    for (unsigned long i = 0; i < mEntities.size(); i++)
         delete mEntities[i];
     mEntities.clear();
 
-    for (unsigned int i = 0; i < mModels.size(); i++)
+    for (unsigned long i = 0; i < mModels.size(); i++)
         delete mModels[i];
     mModels.clear();
 
-    for (unsigned int i = 0; i < mMeshes.size(); i++)
+    for (unsigned long i = 0; i < mMeshes.size(); i++)
         delete mMeshes[i];
     mMeshes.clear();
 }
